pyproject.toml
[tool.poetry]
name = "Jargon Discord"
version = "0.1.0"
description = "Discord Bot Services"
authors = [
    "Erik Karlsson <edkarlsson@users.noreply.github.com>",
    "Armaan <ak2k2@users.noreply.github.com>"
]
readme = "README.md"
repository = "https://github.com/JargonBots/jargon-discord"

[tool.poetry.dependencies]
python = "^3.10"
python-dotenv = ">=0.21.0,<0.22.0"
toml = "^0.10.2"
flask = "^2.2.2"
discord-py = ">=2.1.0,<2.2.0"
openai = ">=0.25.0,<0.26.0"

[tool.poetry.group.dev.dependencies]
autoflake = "^2.0.0"
black = "^22.12.0"
isort = "^5.11.4"
pre-commit = "^2.20.0"
autopep8 = "^2.0.1"
flake8 = "^6.0.0"


[tool.poetry.group.test.dependencies]
pytest = "^7.2.0"


[tool.poetry.group.docs.dependencies]
mkdocs = "^1.4.2"


[tool.poetry.group.sandbox.dependencies]
twilio = "^7.16.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
